Start Driving Out Current Forecast Response
The /forecast endpoint will eventually respond with a JSON representation of
the forecast for a given location. Here, the degenerate case is implemented,
with the forecast for Cupertino provided in the absence of a search parameter.
The simplest working implementation is a hard-coded response, which will be
replaced in subsequent test-driven iterations.

Additionally, the other forecast cases are enumerated as pending request-spec cases.