Handle Invalid Forecast Requests
Implements handling of forecast requests with invalid queries. To keep the
controller's responsibilities as minimal as possible -- just a broker for
request and response data -- the query is passed to a ForecastLookup service
object, which encapsulates the query-validation logic. Thus isolated, the logic
can be extended with real functionality (i.e. whether the query params can be
coerced to latitude and longitude coordinates).