First Pass on Geocoding Query Param
Normalizing the query param to a set of latitude and longitude coordinates will
simplify the process of retrieving the current weather forecast. For this, a
geocoder is necessary. A very nice geocoder for Ruby is the aptly named
"geocoder" gem, which can be configured to perform forward- and
reverse-geocoding operations via numerous APIs. For the purposes of this
assessment, I've opted for the free and fair-use Photon API, which itself takes
advantage of the OpenStreetMap database. Photon's API is not suitable for heavy
load, but the tradeoffs are tolerable for this assessment; and it can be switched
out trivially for a paid, high-availability option, or a self-hosted geocoding
microservice.