Updates Geocoding Logic
Performing a lookup exclusively by zipcode (e.g. "90210") is fairly
unreliable—it often returns coordinates off the coast of Somalia. I've kludged
together a solution that scopes the searches to the United States, which has
the upside of more reliable zipcode lookups, but other downsides. It feels
out-of-scope to revisit the address-lookup logic, and indeed it would be a
non-trivial lift to be more permissive with queries.

I've also extracted Location and NullLocation value objects, to reduce the
responsibilities of the FindLocation service object. They should be extracted
to their own places in the project filesystem at some point.